#42618 publish logging not propagating to console
This change is primarily to help with the scenario where shotgun engine will
report all the progress update coming from the publisher once you have closed
the window.

On a more general level, it's strange that the publish logging is being
dispatched both to the engine console and to the internal console of the
publisher.

This change changes two things:

* The logging relating to publishing (from hooks and processing) - the log
messages shown in the publish console - are no longer propagating into the main
sgtk log namespace.

* The log file button is changed to a copy to clipboard
which allows a user to quickly copy publish progress info into a text editor
(or support ticket). This makes it easier to separate out the publish related
stuff from the copious amounts of debug logging that is also in the log files.

Closes #18.