For #43545: Checks for failed or zero-size thumbnails.
This keeps us from plowing ahead and asking the SG API to upload a non existent or zero-size jpg file, which would cause it to fail ungracefully. Instead, we log the result and continue on without a thumbnail if a problem occurs and one can't be properly written to disk.