#43792 Fix odd behavior of Publisher Tree Widget drag operation. (#23)
PublishTreeWidget rejects drag/drop operations if the selection contains items that are not TopLevelTreeNodeItem. As a result if user shift-selects assets and underlying actions in the publish view, the drag/drop operations fails because of the selected actions. Fix: Modified the behavior of drag operation to consider only elements derived from TopLevelTreeNodeItem, ignoring items of other types.