SG-3055: WIP Adds new methods to graph, adds concept of persistence
Adds the concept of persistent items in the graph. This is used for items that
should remain after the graph is cleared (re-collecting the state). Examples
include files dragged/dropped or added to the graph manually. Items can be
set to persistent via a new property on the items themselves.

Adds the following methods to the graph instance:
`clear()` - Resets the state of the graph by clearing everything except
    persistent nodes. Persistent nodes can be cleared as well via optional
    kwarg.
`get_ancestors()` - Returns all children for a given item recursively
`pformat()` - Returns a pretty formatted string representation of the graph.
    This is useful for debugging
`pprint()` - Convenience method to print the string returned by `pformat()`
`persistent_items` - Convenience property to return all persistent items stored
    in the graph.

Misc. tweaks to logic and docstrings.