Added support for Nuke publishing tree serialization
When publishing using farm_wrapper and serializing the publish tree in Nuke, the presence of a nuke node (eg the writenode) causes the tree.save method to error. Whilst this may be pointing to another greater bug in the code ( `super(_PublishTreeEncoder).default(data)` errors saying the parent class has no attribute `default`) a workaround is to gracefully handle writenodes before the default return is called. 

It will probably be necessary to catch other publish types in this code, and if so, perhaps it should be exposed as a hook so studios won't need to fork this code to define how custom items are serialized. 

This workaround checks that we're in a nuke environment before importing nuke and checking the data is an instance of a nuke.Gizmo. IF it is a gizmo, we serialise the name of the node.