SG-14693 serialize tree bug fix (#112)
Fixes the tree.save method call to super.

Previously it was not passing self, which means unserializable objects were throwing the wrong error. Now it should call the base method and return the standard error for objects that can't be serialized.

Added a test to ensure that an unserializable tree returns the correct error.