SG-9940 description inheritance fix. (#125)
Reworks the description inheritance, so that it is correctly inherited by child item no matter at which level in the hierarchy you enter a description. *Previously the description would only be inherited if it was set in the summary.

*This was a regression that was introduced, previous to that the description was inherited when set on individual items.

The UI has also been updated to make it easier to see what is inheriting the description and providing a link to the item that the description is inherited from.
Setting a description higher up the hierarchy does not override items further down if they are not inheriting the description.